completed all routes

+++ routes/courses.js
@@ -1,6 +1,8 @@
 const Router = require('express').Router();
 const { body, validationResult } = require('express-validator');
 const pool = require('../db');
+const { verify } = require('jsonwebtoken');
+const verifyAdmin = require('../middleware/verifyAdmin');
 
 Router
     // Get route using pagination
@@ -62,8 +64,8 @@ Router
         }
     })
 
-    // Upload all the courses at once
-    .post('/all', async (req, res) => {
+    // Admin Privileges Route
+    .post('/all', verifyAdmin, async (req, res) => {
         try {
             const courses = req.body;
             for (const course of courses) {
@@ -79,11 +81,13 @@ Router
         }
     })
 
-    // Admin Privileges Route
-    .get('/:id', async (req, res) => {
+    .get('/:id', verifyAdmin, async (req, res) => {
         try {
             const result = await pool.query("SELECT * FROM courses WHERE id=$1", [req.params.id]);
+
+            // Checking if the course exists
             if (result.rows.length === 0) return res.status(404).send("No course found with this id");
+
             res.status(200).json(result.rows[0]);
         } catch (error) {
             res.status(500).send(error.message || "Error occurred while fetching course");
@@ -99,9 +103,12 @@ Router
         body('instructor', "Instructor cannot be empty").notEmpty(),
         body('description', "Description cannot be empty").notEmpty(),
         body('price', "Price cannot be empty").notEmpty(),
-    ], async (req, res) => {
+    ], verifyAdmin, async (req, res) => {
+
+        // Checking the condition for all fields
         const errors = validationResult(req);
         if (!errors.isEmpty()) return res.status(400).json({ errors: errors.array() });
+
         try {
             const result = await pool.query("INSERT INTO courses (name, category, level, popularity, duration, instructor, description, price) VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING *", [...Object.values(req.body)]);
             res.status(200).json(result.rows[0]);
@@ -119,24 +126,38 @@ Router
         body('instructor', "Instructor cannot be empty").notEmpty().optional(),
         body('description', "Description cannot be empty").notEmpty().optional(),
         body('price', "Price cannot be empty").notEmpty().optional(),
-    ], async (req, res) => {
+    ], verifyAdmin, async (req, res) => {
+        // Checking the condition for all fields
         const errors = validationResult(req);
         if (!errors.isEmpty()) return res.status(400).json({ errors: errors.array() });
-        if (Object.keys(req.body).length === 0) return res.status(400).send("Atleast 1 field is requried");
+
+        // Checking if atleast 1 field is present
+        if (Object.keys(req.body).length === 0) return res.status(400).send("Atleast 1 field is required");
 
         try {
-            const updateQuery = "UPDATE courses SET " + Object.keys(req.body).map((key, index) => `${key}=$${index + 1}`).join(', ') + " WHERE id=$" + (Object.keys(req.body).length + 1) + ' RETURNING *';
-            console.log(updateQuery)
-            const result = await pool.query(updateQuery, [...Object.values(req.body), req.params.id]);
-            res.status(200).json(result.rows[0]);
 
+            // Checking if the course exists
+            const course = await pool.query("SELECT * FROM courses WHERE id=$1", [req.params.id]);
+            if (course.rows.length === 0) return res.status(404).send("No course exists with this id");
+            
+            // Query for updating the course according to the fields provided
+            const updateQuery = "UPDATE courses SET " + Object.keys(req.body).map((key, index) => `${key}=$${index + 1}`).join(', ') + " WHERE id=$" + (Object.keys(req.body).length + 1) + ' RETURNING *';
+            
+            const result = await pool.query(updateQuery, [...Object.values(req.body), req.params.id]);
+            
+            res.status(200).json(result.rows[0]);
+            
         } catch (error) {
             res.status(500).send(error.message || "Error occurred while updating course");
         }
     })
-
-    .delete('/:id', async (req, res) => {
+    
+    .delete('/:id', verifyAdmin, async (req, res) => {
         try {
+            // Checking if the course exists
+            const course = await pool.query("SELECT * FROM courses WHERE id=$1", [req.params.id]);
+            if (course.rows.length === 0) return res.status(404).send("No course exists with this id");
+            
             await pool.query("DELETE FROM courses WHERE id=$1", [req.params.id]);
             res.status(200).send("Course Deleted successfully");
         } catch (error) {
